requirements.txt
apispec==5.1.1
attrs==21.4.0
click==8.0.3
flasgger==0.9.5
Flask==2.0.2
flask-apispec==0.11.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.0.1
marshmallow==3.14.1
mistune==2.0.2
packaging==21.3
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
webargs==8.1.0
Werkzeug==2.0.2
zipp==3.7.0
